requirements.txt
# Core Django
Django>=5.0,<5.1
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
python-decouple>=3.8
django-unfold>=0.19.0

# Database
psycopg2-binary>=2.9.9

# API Documentation
drf-spectacular>=0.27.0

# Authentication
djangorestframework-simplejwt>=5.3.1

# Caching & Task Queue
redis>=5.0.0
celery>=5.3.4
django-redis>=5.4.0

# Rate Limiting
django-ratelimit>=4.1.0

# Utilities
Pillow>=10.1.0
python-dateutil>=2.8.2
requests>=2.31.0

# Development
django-extensions>=3.2.3

# Production Dependencies
gunicorn>=21.2.0
whitenoise>=6.6.0
psutil>=5.9.0
sentry-sdk[django]>=1.38.0
flower>=2.0.1

